requirements-api.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.26.4
transformers==4.55.4
onnxruntime-gpu==1.22.0